@@ -21,7 +21,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
 
         <springdoc-openapi.version>1.5.10</springdoc-openapi.version>
         <velocity.version>2.3</velocity.version>
